added conditional test

+++ itmt-495-595/professors-project-code/creation-scripts/delete-original-box-files.ps1
@@ -1,5 +1,10 @@
 Set-Location -path ../../build
 
-Remove-Item -Verbose *.box
+if (Test-Path -Path ./*.box -PathType Leaf) {
+    Remove-Item -Verbose *.box
+}
+else {
+    Write-Host "No .box files to delete"
+}
 
-Set-Location -path ../../professors-project-code/creation-scripts
+Set-Location -path ../professors-project-code/creation-scripts
